Fix crash when switching between bricklets.
Previously a device_information object, that could contain
stale data, was given to the plugin started by selecting another
device in the main drop down list.

Now the device_information is not reused, instead an explicit enumerate
is triggered when selecting another device. This will in turn fill
the device_information with fresh data and trigger a plugin's new_enum
method.