esp-flash: Fix flash config data patch
This broken check resulted in the web interface of the WIFI Extension 2.0
not working and just reporting 404.

The original check was fine in Python 2 where the b prefix on a str is
ignored and the index operation yields a str. In Python 3 image is not
an str but a bytes object. The index operation on a bytes object yields
an int that always compares as False to a bytes object.